Keep chart-legend on larger resolutions for IE
IE sucks and does not render foreignObjects thus we should
keep the legend because the axis labels won't be shown.

Also the chart height was adjusted for smaller resolutions, so the
legend would overlap the axis label text. This is necessary because
display:none and visibility:collapse is not working for the axis label.